# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BSOPT API",
    "version": "2.1.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get Current User Profile",
        "description": "Get current user's profile information.",
        "operationId": "get_current_user_profile_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete Current User Account",
        "description": "Delete current user's account.\n\nThis is a soft delete - marks account as inactive.",
        "operationId": "delete_current_user_account_api_v1_users_me_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Update Current User Profile",
        "description": "Update current user's profile.",
        "operationId": "update_current_user_profile_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/me/stats": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User Stats",
        "description": "Get usage statistics and rate limit information for the current user.",
        "operationId": "get_user_stats_api_v1_users_me_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserStatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get User By Id",
        "description": "Get user by ID (admin only).",
        "operationId": "get_user_by_id_api_v1_users__user_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SuccessResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Success message"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "SuccessResponse",
        "description": "Standard success response.",
        "example": {
          "data": {
            "id": "123"
          },
          "message": "Operation completed successfully",
          "success": true
        }
      },
...
      "UserResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "title": "Email",
            "description": "User email"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "description": "User's full name"
          },
          "tier": {
            "type": "string",
            "title": "Tier",
            "description": "Subscription tier"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "Account active status"
          },
          "is_verified": {
            "type": "boolean",
            "title": "Is Verified",
            "description": "Email verified status"
          },
          "is_mfa_enabled": {
            "type": "boolean",
            "title": "Is Mfa Enabled",
            "description": "MFA enabled status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "Account creation date"
          },
          "last_login": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Login",
            "description": "Last login timestamp"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "tier",
          "is_active",
          "is_verified",
          "is_mfa_enabled",
          "created_at"
        ],
        "title": "UserResponse",
        "description": "User profile response.",
        "example": {
          "created_at": "2024-01-01T00:00:00Z",
          "email": "user@example.com",
          "full_name": "John Doe",
          "id": "123e4567-e89b-12d3-a456-426614174000",
          "is_active": true,
          "is_mfa_enabled": false,
          "is_verified": true,
          "last_login": "2024-01-15T10:30:00Z",
          "tier": "pro"
        }
      },
      "UserStatsResponse": {
        "properties": {
          "total_requests": {
            "type": "integer",
            "title": "Total Requests",
            "description": "Total API requests"
          },
          "requests_today": {
            "type": "integer",
            "title": "Requests Today",
            "description": "Requests made today"
          },
          "requests_this_month": {
            "type": "integer",
            "title": "Requests This Month",
            "description": "Requests made this month"
          },
          "rate_limit_remaining": {
            "type": "integer",
            "title": "Rate Limit Remaining",
            "description": "Remaining requests in current window"
          },
          "rate_limit_reset": {
            "type": "string",
            "format": "date-time",
            "title": "Rate Limit Reset",
            "description": "When rate limit resets"
          }
        },
        "type": "object",
        "required": [
          "total_requests",
          "requests_today",
          "requests_this_month",
          "rate_limit_remaining",
          "rate_limit_reset"
        ],
        "title": "UserStatsResponse",
        "description": "User statistics response.",
        "example": {
          "rate_limit_remaining": 750,
          "rate_limit_reset": "2024-01-15T11:00:00Z",
          "requests_this_month": 3500,
          "requests_today": 250,
          "total_requests": 15000
        }
      },
      "UserUpdateRequest": {
        "properties": {
          "full_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name",
            "description": "User's full name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "New email address"
          }
        },
        "type": "object",
        "title": "UserUpdateRequest",
        "description": "User profile update request.",
        "example": {
          "full_name": "John Smith"
        }
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}